triggerActionDefs.js moved out of synthConfig...
Trigger actions are defined on the host, not on the synth.
They should be defined differently for different hosts.
The current definitions reference particular select controls.
The AssistantPerformer could use ActionDefs containing:
a) the key that triggers the action
b) an actionStack (that includes noAction) that is popped each time the key is pressed
c) actions could include
  - noAction,
  - preset changes (including mixture changes),
  - tuning changes,
  - whole sequences of NoteOn messsages (in WebWorkers)