Removed MIDI REGISTERED_PARAMETER functions...
This simplifies things considerably.
updatePitchWheelSensitivity(channel, semitones) is now called directly by the host, in the same way that setTuning(...) and setCentsOfffset(...) are currently realized.
If these public functions are to be called from the AssistantPerformer, I will have to find a non-MIDI way to do it. I think this could be fairly easy to do by exposing an array of synth functions to the host.