Moved stateDefs to host as presets...
Removed synth setChannelState code and the SET_CHANNEL_STATE control type.
host now has a button for exporting the current host state as a preset.
The resulting code can be added to presets.js.
TODO: Add a preset selector to host. This will be set up with options corresponding to the presets in presets.js