Disabled triggers and selection of predefined settings while recording. selection (1).
The current trigger key works like a normal midi key while recording.
The Settings Select is a place to store preselected settings that may be useful in future.
The trigger key is a way to cycle through settings presets, and a way to reactivate the settingsSelect when it is currently disabled. (It is disabled when the current settings are not the same as the settings shown on the settingsSelect.