Implemented velocityPitchChange control... (TODO)
TODO: Finalize the layout of the velocityPitchSensitivityNumberInput in the Notes section in the host's GUI.

Each note's velocity can now affect its sounding pitch, whereby all octaves are exact for each absolute pitch.
The velocityPitchSensitivity is a value (like pitchWheelSensitivity) (in range 0..0.6 in the current host, but could be maximum of 0..1) set by a new velocityPitchSensitivityNumberInput in the GUI.
To implement exact, sounding octaves: the channelControl object in the ResidentSynth now has a velocityPitchValue14Bit array, with one entry per absolute pitch. If the value is undefined, the note's calculated velocityPitchValue14Bit value is set here and used by subsequent equal absolute pitches until none of them is sounding any more.