Revised velocityPitchSensitivity (max deviation is now value/10 semitones)...
Also simplified the code.
The maximum pitch deviation (at velocity 127) is now the control's value/10 semitones. In other words, in equal temperament tuning: If this value is 0, the velocity will have no effect on the pitch; if this value is 10, velocity 127 will raise the pitch by 1 semitone; if this value is 127, velocity 127 will raise the pitch by 12.7 semitones.