New implementation of channel settings...
Minimizing the number of messages that are sent to the synth when calling onSettingsSelect().
Settings should override existing settings. also in the sequence of settings in the settingsSelect, so that the trigger works on the select index.
Note that if the settingSelect is changed manually, it should call all the settings changes in order from the beginning of its list.
This code is still being debugged.