Added document.onkeydown(event) handler onTriggerKeyDown(event)... see TODO
This is going to replace the existing trigger key control and function. It converts the text keyboard into a bank of control buttons, that can in principle be used to trigger any kind of macro or script.
The current onTriggerKeyDown(event) function deals with the following keys:
<space> increments/rotates the settingsSelect.selectedIndex, making the current triggerKey redundant.
The '0'..'9' keys select settingIndex 0..9 respectively, (incrementing/rotating if out of range).
The 'a'..'z' keys select settingIndex 10..35 respectively (incrementing/rotating if out of range).
TODO: remove the existing triggerKey input control, and all its related code completely! This includes removing the triggerKeys entry in the ResSynth.synthSettingsDefs!