Added Equal Temperament, exact fifths, wide octaves tuning...
Having 7 equal steps between each perfect fifth -- pitch factor between keys on keyboard is (3/2)^(1/7). So all fifths are perfect. Most noticeable is that the major third is better than in 12-tone equal temperament with exact octaves. The octaves are a bit wierd, but fifths can be more important, depending on how the harmony is designed...