Added new PERFECT_KEYBOARD_INTERVAL tuning group with 13 tunings...
Each tuning is defined by the ratio of frequencies allocated to adjacent keys on the standard keyboard. The frequencies are defined so that a particular interval is "perfect" -- i.e. is exactly related to an odd harmonic. The tunings are such that each fingerable interval in turn is the basis for the tuning. The interval is then divided equally into the requisite number of adjacent keys. For example: a perfect fifth is the relatio 3/2, comprising 7 adjacent keys. So the frequency ratio between adjacent keys is (3/2)^(1/7).